backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
supabase==2.3.5
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# AI & Multi-Agent
crewai==0.11.0
groq==0.4.2
openai==1.12.0
langchain-community==0.0.19
duckduckgo-search==4.3.1
google-search-results==2.4.2
httpx==0.26.0

# Vector Database
qdrant-client==1.7.3

# Authentication
# clerk-backend-api==0.2.0 (Removed, replaced with NextAuth verification)
python-jose==3.3.0

# WebSocket
python-socketio==5.11.1
websockets==12.0

# Utilities
python-multipart==0.0.7
aiofiles==23.2.1
tenacity==8.2.3

# Monitoring
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
